rename `array_tuple` as `arrays_zip`
