+++ src/query/functions/tests/it/scalars/testdata/array.txt
@@ -2382,3 +2382,38 @@ output domain  : [{0.0..=5.6} ∪ {NULL}]
 output         : [5.6, 3.4, 2.2, 1.2, NULL, NULL]
 
 
+ast            : arrays_zip(NULL, NULL)
+raw expr       : arrays_zip(NULL, NULL)
+checked expr   : arrays_zip<NULL, NULL>(NULL, NULL)
+optimized expr : [(NULL, NULL)]
+output type    : Array(Tuple(NULL, NULL))
+output domain  : [({NULL}, {NULL})]
+output         : [(NULL, NULL)]
+
+
+ast            : arrays_zip(1, 2, 'a')
+raw expr       : arrays_zip(1, 2, 'a')
+checked expr   : arrays_zip<UInt8, UInt8, String>(1_u8, 2_u8, "a")
+optimized expr : [(1, 2, 'a')]
+output type    : Array(Tuple(UInt8, UInt8, String))
+output domain  : [({1..=1}, {2..=2}, {"a"..="a"})]
+output         : [(1, 2, 'a')]
+
+
+ast            : arrays_zip([1,2,3], ['a','b','c'], 10)
+raw expr       : arrays_zip(array(1, 2, 3), array('a', 'b', 'c'), 10)
+checked expr   : arrays_zip<Array(UInt8), Array(String), UInt8>(array<T0=UInt8><T0, T0, T0>(1_u8, 2_u8, 3_u8), array<T0=String><T0, T0, T0>("a", "b", "c"), 10_u8)
+optimized expr : [(1, 'a', 10), (2, 'b', 10), (3, 'c', 10)]
+output type    : Array(Tuple(UInt8, String, UInt8))
+output domain  : [({1..=3}, {"a"..="c"}, {10..=10})]
+output         : [(1, 'a', 10), (2, 'b', 10), (3, 'c', 10)]
+
+
+error: 
+  --> SQL:1:1
+  |
+1 | arrays_zip([1,2,3], ['a','b'], 10)
+  | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ array length must be equal, but got 3 and 2 while evaluating function `arrays_zip([1, 2, 3], ['a', 'b'], 10)` in expr `arrays_zip(array(1, 2, 3), array('a', 'b'), 10)`
+
+
+
