+++ tests/sqllogictests/suites/query/functions/02_0015_function_tuples.test
@@ -6,15 +6,3 @@ SELECT tuple([1,2,3], ['a','b','c'], 10);
 ----
 ([1,2,3],['a','b','c'],10)
 
-query T
-SELECT array_tuple([1,2,3], ['a','b','c'], 10);
-----
-[(1,'a',10),(2,'b',10),(3,'c',10)]
-
-statement error 1006
-SELECT array_tuple([1,2,3], ['a','b'], 10);
-
-query T
-SELECT array_tuple(1, 'a', null);
-----
-[(1,'a',NULL)]
